+++ packages/plugin-koa/src/koa.js
@@ -15,8 +15,8 @@ module.exports = {
   load: client => {
     const requestHandler = async (ctx, next) => {
       // Get a client to be scoped to this request. If sessions are enabled, use the
-      // startSession() call to get a session client, otherwise, clone the existing client.
-      const requestClient = client._config.autoTrackSessions ? client.startSession() : clone(client)
+      // resumeSession() call to get a session client, otherwise, clone the existing client.
+      const requestClient = client._config.autoTrackSessions ? client.resumeSession() : clone(client)
 
       ctx.bugsnag = requestClient
 
@@ -47,8 +47,8 @@ module.exports = {
 
     requestHandler.v1 = function * (next) {
       // Get a client to be scoped to this request. If sessions are enabled, use the
-      // startSession() call to get a session client, otherwise, clone the existing client.
-      const requestClient = client._config.autoTrackSessions ? client.startSession() : clone(client)
+      // resumeSession() call to get a session client, otherwise, clone the existing client.
+      const requestClient = client._config.autoTrackSessions ? client.resumeSession() : clone(client)
 
       this.bugsnag = requestClient