+++ packages/plugin-restify/src/restify.js
@@ -17,8 +17,8 @@ module.exports = {
       const dom = domain.create()
 
       // Get a client to be scoped to this request. If sessions are enabled, use the
-      // startSession() call to get a session client, otherwise, clone the existing client.
-      const requestClient = client._config.autoTrackSessions ? client.startSession() : clone(client)
+      // resumeSession() call to get a session client, otherwise, clone the existing client.
+      const requestClient = client._config.autoTrackSessions ? client.resumeSession() : clone(client)
 
       // attach it to the request
       req.bugsnag = requestClient